finished all integration tests
